Split out Axis and Dim
