+++ src/dimension/mod.rs
@@ -5,7 +5,6 @@
 // <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
-use std::cmp::Ordering;
 use std::fmt::Debug;
 use std::ops::{Index, IndexMut};
 
@@ -18,6 +17,12 @@ use error::{from_kind, ErrorKind, ShapeError};
 use ZipExt;
 use {Ix0, Ix1, Ix2, Ix3, Ix4, Ix5, Ix6, IxDyn};
 use {ArrayView1, ArrayViewMut1};
+
+pub use self::dim::*;
+pub use self::axis::Axis;
+
+pub mod dim;
+mod axis;
 
 /// Calculate offset from `Ix` stride converting sign properly
 #[inline(always)]
@@ -1241,193 +1246,3 @@ mod test {
     }
 }
 
-/// An axis index.
-///
-/// An axis one of an array’s “dimensions”; an *n*-dimensional array has *n* axes.
-/// Axis *0* is the array’s outermost axis and *n*-1 is the innermost.
-///
-/// All array axis arguments use this type to make the code easier to write
-/// correctly and easier to understand.
-#[derive(Copy, Eq, Ord, Hash, Debug)]
-pub struct Axis(pub usize);
-
-impl Axis {
-    #[inline(always)]
-    pub fn axis(&self) -> usize { self.0 }
-}
-
-macro_rules! clone_from_copy {
-    ($typename:ident) => {
-        impl Clone for $typename {
-            #[inline]
-            fn clone(&self) -> Self { *self }
-        }
-    }
-}
-
-macro_rules! derive_cmp {
-    ($traitname:ident for $typename:ident, $method:ident -> $ret:ty) => {
-        impl $traitname for $typename {
-            #[inline(always)]
-            fn $method(&self, rhs: &Self) -> $ret {
-                (self.0).$method(&rhs.0)
-            }
-        }
-    }
-}
-
-derive_cmp!{PartialEq for Axis, eq -> bool}
-derive_cmp!{PartialOrd for Axis, partial_cmp -> Option<Ordering>}
-clone_from_copy!{Axis}
-
-pub trait DimPrivate<I> {
-    fn new(index: I) -> Self;
-    fn ix(&self) -> &I;
-    fn ixm(&mut self) -> &mut I;
-}
-
-pub use self::dim::*;
-pub mod dim {
-    use super::IntoDimension;
-    use super::DimPrivate;
-    use super::Dimension;
-    use Ix;
-
-    use itertools::zip;
-
-    /// Dimension description.
-    ///
-    /// `Dim` describes the number of axes and the length of each axis
-    /// in an array. It is also used as an index type.
-    #[derive(Copy, Clone, Debug, PartialEq, Eq, Default)]
-    pub struct Dim<I: ?Sized> {
-        index: I,
-    }
-
-    impl<I> DimPrivate<I> for Dim<I> {
-        fn new(index: I) -> Dim<I> {
-            Dim {
-                index: index,
-            }
-        }
-        fn ix(&self) -> &I { &self.index }
-        fn ixm(&mut self) -> &mut I { &mut self.index }
-    }
-
-    /// Create a new dimension value.
-    #[allow(non_snake_case)]
-    pub fn Dim<T>(index: T) -> T::Dim
-        where T: IntoDimension
-    {
-        index.into_dimension()
-    }
-
-    impl<I: ?Sized> PartialEq<I> for Dim<I>
-        where I: PartialEq,
-    {
-        fn eq(&self, rhs: &I) -> bool {
-            self.index == *rhs
-        }
-    }
-
-    use std::ops::{Add, Sub, Mul, AddAssign, SubAssign, MulAssign};
-
-    macro_rules! impl_op {
-        ($op:ident, $op_m:ident, $opassign:ident, $opassign_m:ident, $expr:ident) => {
-        impl<I> $op for Dim<I>
-            where Dim<I>: Dimension,
-        {
-            type Output = Self;
-            fn $op_m(mut self, rhs: Self) -> Self {
-                $expr!(self, &rhs);
-                self
-            }
-        }
-
-        impl<I> $opassign for Dim<I>
-            where Dim<I>: Dimension,
-        {
-            fn $opassign_m(&mut self, rhs: Self) {
-                $expr!(*self, &rhs);
-            }
-        }
-
-        impl<'a, I> $opassign<&'a Dim<I>> for Dim<I>
-            where Dim<I>: Dimension,
-        {
-            fn $opassign_m(&mut self, rhs: &Self) {
-                for (x, &y) in zip(self.slice_mut(), rhs.slice()) {
-                    $expr!(*x, y);
-                }
-            }
-        }
-
-        }
-    }
-
-    macro_rules! impl_single_op {
-        ($op:ident, $op_m:ident, $opassign:ident, $opassign_m:ident, $expr:ident) => {
-        impl $op<Ix> for Dim<[Ix; 1]>
-        {
-            type Output = Self;
-            #[inline]
-            fn $op_m(mut self, rhs: Ix) -> Self {
-                $expr!(self, rhs);
-                self
-            }
-        }
-
-        impl $opassign<Ix> for Dim<[Ix; 1]> {
-            #[inline]
-            fn $opassign_m(&mut self, rhs: Ix) {
-                $expr!((*self)[0], rhs);
-            }
-        }
-        };
-    }
-
-    macro_rules! impl_scalar_op {
-        ($op:ident, $op_m:ident, $opassign:ident, $opassign_m:ident, $expr:ident) => {
-        impl<I> $op<Ix> for Dim<I>
-            where Dim<I>: Dimension,
-        {
-            type Output = Self;
-            fn $op_m(mut self, rhs: Ix) -> Self {
-                $expr!(self, rhs);
-                self
-            }
-        }
-
-        impl<I> $opassign<Ix> for Dim<I>
-            where Dim<I>: Dimension,
-        {
-            fn $opassign_m(&mut self, rhs: Ix) {
-                for x in self.slice_mut() {
-                    $expr!(*x, rhs);
-                }
-            }
-        }
-        };
-    }
-
-    macro_rules! add {
-        ($x:expr, $y:expr) => { $x += $y; }
-    }
-    macro_rules! sub {
-        ($x:expr, $y:expr) => { $x -= $y; }
-    }
-    macro_rules! mul {
-        ($x:expr, $y:expr) => { $x *= $y; }
-    }
-    impl_op!(Add, add, AddAssign, add_assign, add);
-    impl_single_op!(Add, add, AddAssign, add_assign, add);
-    impl_op!(Sub, sub, SubAssign, sub_assign, sub);
-    impl_single_op!(Sub, sub, SubAssign, sub_assign, sub);
-    impl_op!(Mul, mul, MulAssign, mul_assign, mul);
-    impl_scalar_op!(Mul, mul, MulAssign, mul_assign, mul);
-}
-
-
-#[test]
-fn test_dim() {
-}
